refactor + feature: launch file auto-alias and split up launch files
When the package is compiled, it will automatically make transparant
aliases for every launch file in the package. This avoids loosing track
of all existing aliases that are kept in tue-env. Furthermore, I removed
the middleware.launch and combined the challenge with the specific
requried middleware components. For testing and demos, a free_mode
launch file is created. This way, at the real robot, we can autostart
the free_mode.launch file, with amigo-free-mode and all robot consoles
etc will work. If we would like to work with challenges, we can stop
this service and use only one launch file to start the challenge with
its specific dependencies: amigo-challenge-r5cop for example. Note that
all aliases are now prefixed with the bringup name (amigo) followed by
the launch file name. So astart and amiddle do not exist anymore,
instead we have: amigo-start, amigo-free-mode, amigo-rviz,
amigo-challenge-r5cop, amigo-challenge-XX etc.